feat: :sparkles: add error for delete

+++ src/services/product.service.ts
@@ -1,7 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Error, Model } from 'mongoose';
 import { ProductDTO } from 'src/dto/product.dto';
+import { IProduct } from 'src/models/product.model';
 import { Product } from 'src/schemas/product.schema';
 
 @Injectable()
@@ -18,7 +23,25 @@ export default class ProductService {
     return newProduct.save();
   }
 
-  async deleteProduct(idProduct: string): Promise<Product> {
-    return this.productModel.findByIdAndDelete(idProduct);
+  async deleteProduct(idProduct: IProduct['id']): Promise<Product> {
+    try {
+      if (!idProduct) {
+        throw new BadRequestException('Product ID is required');
+      }
+
+      const deleteProduct =
+        await this.productModel.findByIdAndDelete(idProduct);
+
+      if (!deleteProduct) {
+        throw new NotFoundException('Product ID Not Found');
+      }
+
+      return deleteProduct;
+    } catch (error) {
+      if (error instanceof Error.CastError) {
+        throw new BadRequestException('Invalid ID format');
+      }
+      throw error; // Re-throw other unexpected errors
+    }
   }
 }
